% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_usps.R
\name{track_usps}
\alias{track_usps}
\title{track_usps}
\usage{
track_usps(number)
}
\arguments{
\item{number}{tracking number as numeric or string}
}
\value{
string containing status of package
}
\description{
track_usps
}
\examples{
track_usps(3505504435669285650242)
}
